problem: really did need to do different things for v4 vs v6.
Add back in the functionality to do different things for v4 addresses vs
v6 addresses.  Since it is now batching announcements, I need to group
announcements by v4 or v6 before executing the template.  This ensures
that the list of addresses passed to the block function is all v4 or all
v6.