process/template: delegate more control over route generation to template
The previous implementation of make_routes was built with an assumption
that exactly one exabgp API command would be rendered based on any list
of cidrs.  This assumption makes sense for blocking via null routes + RPF
to cover blocking in forward and reverse directions.

The single-route-per-block assumption does not hold, however, for
scenarios where the blocking is accomplished via BGP flowspec routes.  In
this case, two exabgp route announcements per cidr would need to be added.
One for forward traffic and the other for reverse traffic flows.
Something like this for each cidr:
  * announce flow source      198.51.100.0/24 extended-community rate-limit:0
  * announce flow destination 198.51.100.0/24 extended-community rate-limit:0

This level of control requires that the template is able to loop over the
cidrs as well as render the action (announce/withdraw) for each line as
appropriate.

This change gives the template full control by allowing the template to
render the action as well as to receive the cidrs as a list rather than a
space-separated string.